+++ R/src-mysql.r
@@ -121,7 +121,7 @@ src_desc.src_mysql <- function(x) {
 }
 
 #' @export
-src_translate_env.src_mysql <- function(x) {
+sql_translate_env.MySQLConnection <- function(x) {
   sql_variant(
     base_scalar,
     sql_translator(.parent = base_agg,