+++ R/translate-sql-window.r
@@ -7,7 +7,7 @@
 # translate_window_where(quote(n() > 10), players)
 # translate_window_where(quote(rank() > cumsum(AB)), players)
 # translate_window_where(list(quote(x == 1), quote(n() > 2)), players)
-translate_window_where <- function(expr, tbl, con = NULL) {
+translate_window_where <- function(expr, vars, con = NULL) {
   # Simplest base case: atomic vector or name ---------------------------------
   if (is.atomic(expr) || is.name(expr)) {
     return(list(
@@ -17,12 +17,12 @@ translate_window_where <- function(expr, tbl, con = NULL) {
   }
 
   # Other base case is an aggregation function --------------------------------
-  variant <- src_translate_env(tbl)
+  variant <- sql_translate_env(con)
   agg_f <- ls(envir = variant$window)
 
   if (is.call(expr) && as.character(expr[[1]]) %in% agg_f) {
     name <- unique_name()
-    sql <- translate_sql_q(list(expr), tbl, env = NULL, window = TRUE)
+    sql <- translate_sql_(list(expr), vars = vars)
 
     return(list(
       expr = as.name(name),
@@ -33,12 +33,12 @@ translate_window_where <- function(expr, tbl, con = NULL) {
   # Recursive cases: list and all other functions -----------------------------
 
   if (is.list(expr)) {
-    args <- lapply(expr, translate_window_where, tbl = tbl, con = con)
+    args <- lapply(expr, translate_window_where, vars = vars, con = con)
 
     env <- sql_env(call, variant, con = con)
     sql <- lapply(lapply(args, "[[", "expr"), eval, env = env)
   } else {
-    args <- lapply(expr[-1], translate_window_where, tbl = tbl, con = con)
+    args <- lapply(expr[-1], translate_window_where, vars = vars, con = con)
 
     call <- as.call(c(expr[[1]], lapply(args, "[[", "expr")))
     env <- sql_env(call, variant, con = con)